Created AppleTree Object
Created separate Trunk and Leaves objects and then added them to an empty parent game object called Apple Tree so they would move and scale together. Added materials to recolor to the tree, and then moved the AppleTree object to the prefab folder to create a prefab of the object.