Created Apple Object
Created Apple object and added a new material to color the apple. Added a rigidbody component and left gravity on so the apple object will fall. Also created a new tag called "Apple" for all of the apple objects. Added the apple object to the prefab folder to make it a prefab.